+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Contributed:
 
 Changes:
 
+- Adjust `isBigInt` to not do `instanceof` check
+- Use `@polkadot/x-bigint` for `BigInt` constructor
+- Split `@noble/hashes` fork into `x-noble-hashes` (with above `BigInt`)
 - Ensure consequtive capitals in `stringCamelCase` are lowercased